% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_gcd_vector}
\alias{bench_gcd_vector}
\title{Greatest Common Divisor (Vectorized)}
\usage{
bench_gcd_vector(x, y)
}
\arguments{
\item{x}{First vector of integers}

\item{y}{Second vector of integers}
}
\description{
Computes GCD for pairs of integers using Euclidean algorithm.
}
