% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_grouped_mean}
\alias{bench_grouped_mean}
\title{Grouped Aggregation}
\usage{
bench_grouped_mean(x, groups)
}
\arguments{
\item{x}{data vector}

\item{groups}{Integer vector of group IDs}
}
\description{
Computes mean by group (like tapply or dplyr::summarize).
}
