% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_hilbert_matrix}
\alias{bench_hilbert_matrix}
\title{Hilbert Matrix Creation}
\usage{
bench_hilbert_matrix(n)
}
\arguments{
\item{n}{Matrix dimension}
}
\description{
Creates Hilbert matrix H[i,j] = 1/(i+j-1).
}
