% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_matrix_manip}
\alias{bench_matrix_manip}
\title{Matrix Creation, Transpose and Reshape}
\usage{
bench_matrix_manip(v, nrow, ncol)
}
\arguments{
\item{v}{Input vector}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}
}
\description{
Creates matrix from vector, transposes, reshapes, transposes again.
  Tests matrix manipulation overhead.
}
