% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_matrix_power}
\alias{bench_matrix_power}
\title{Matrix Power}
\usage{
bench_matrix_power(a, exponent)
}
\arguments{
\item{a}{Input matrix}

\item{exponent}{Power to raise to}
}
\description{
Raises each element of matrix to a power.
}
