% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_mean_na_rm}
\alias{bench_mean_na_rm}
\title{NA Handling}
\usage{
bench_mean_na_rm(x)
}
\arguments{
\item{x}{data vector with NAs}
}
\description{
Computes mean excluding NAs, like mean(x, na.rm=TRUE).
}
