% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bench_toeplitz_matrix}
\alias{bench_toeplitz_matrix}
\title{Toeplitz Matrix Creation}
\usage{
bench_toeplitz_matrix(n)
}
\arguments{
\item{n}{Matrix dimension}
}
\description{
Creates Toeplitz matrix T[i,j] = |i-j| + 1 using nested loops.
}
